Update FinalProject.ipynb
Worked on some police and transit graphs at the end of the FinalProject file. Not done yet though. we still need to analyze the correlations between the CPI, salary and housing changes. Visually there seems to be correlations but idk how to actually put a test for that in python